Extracting common template rendering logic into a trait.
